Move tests for saving, copying and deleting a judgment into their own file
